+++ src/main/kotlin/com/zepben/evolve/database/sqlite/cim/upgrade/UpgradeRunner.kt
@@ -10,9 +10,9 @@ package com.zepben.evolve.database.sqlite.cim.upgrade
 
 import com.zepben.evolve.database.paths.DatabaseType
 import com.zepben.evolve.database.sqlite.cim.tables.tableCimVersion
-import com.zepben.evolve.database.sqlite.extensions.configureBatch
-import com.zepben.evolve.database.sqlite.common.TableVersion
 import com.zepben.evolve.database.sqlite.cim.upgrade.changesets.*
+import com.zepben.evolve.database.sqlite.common.TableVersion
+import com.zepben.evolve.database.sqlite.extensions.configureBatch
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 import java.io.IOException
@@ -37,7 +37,8 @@ class UpgradeRunner @JvmOverloads constructor(
     internal val postSplitChangeSets: List<ChangeSet> = listOf(
         changeSet50(),
         changeSet51(),
-        changeSet52()
+        changeSet52(),
+        changeSet53()
     ),
     private val tableVersion: TableVersion = tableCimVersion
 ) {